Refactor indexedDB functions
The setDBList and setDBListItem functions have been separated to have a create and update function for each. Previously, the set functions handled both cases. This allows stricter typing for database object stores. Now the create functions must provide all the properties for the store, whereas the update functions only need provide the ID and the updated property. The rest of the properties will first be fetched from the database. This will also be necessary for when we throttle the update functions while the user is typing.