Reverted custom setter for DBSyncState
The custom setter disallowed updating state if the DBError state was true. However, on resyncDatabase success, the DBerror state and DBSyncState both need to be set to true on the same re-render. This was made impossible with the custom setter. As such, the custom setter has been removed and now the default state setter funciton is passed with the context again as it was before. It will be the event handler function's responsibility to check the DBError state and call the DBSyncState setter conditionally.