Refactor Contexts and Providers
Context is now created and exported from a different file than the Context Provider component. This is done because fast-refresh for hot module reload does not support exporting constants and components from the same file. Renamed "providers" folder to "contexts" for clarity as it now also includes the new Contexts files as well as the Provider componenets.